requirements.txt
python-jobspy==1.1.82
pandas==2.3.3
pydantic==2.12.5
